requirements.txt
# Core dependencies
requests
pyyaml
kubernetes
schedule
loguru

# Optional visualization (if needed)
flask
